refactor(agent/filefinder): merge files, remove dead code, compact (-145 lines)
- Merge normalize.go + score.go into text.go
- Remove unused IncludeHidden field from SearchOptions
- Add copyPostings[K] generic helper in delta.go
- Aggressive blank line removal across all source files
- Compact struct literals and single-line returns

Source: 1333 -> 1188 lines (-10.9%)
Total: 3418 -> 3273 lines
Benchmarks: identical performance, race-clean
