+++ agent/filefinder/delta.go
@@ -11,7 +11,6 @@ const (
 	FlagSymlink FileFlag = 2
 )
 
-// doc is a single indexed document (file or directory).
 type doc struct {
 	path    string
 	baseOff int
@@ -50,31 +49,22 @@ func NewIndex() *Index {
 	}
 }
 
-// Add inserts a path into the index, tombstoning any previous entry. Returns the doc ID.
+// Add inserts a path into the index, tombstoning any previous entry.
 func (idx *Index) Add(path string, flags uint16) uint32 {
 	norm := string(normalizePathBytes([]byte(path)))
-
 	if oldID, ok := idx.byPath[norm]; ok {
 		idx.deleted[oldID] = true
 	}
-
 	id := uint32(len(idx.docs))
 	baseOff, baseLen := extractBasename([]byte(norm))
-
 	idx.docs = append(idx.docs, doc{
-		path:    norm,
-		baseOff: baseOff,
-		baseLen: baseLen,
-		depth:   strings.Count(norm, "/"),
-		flags:   flags,
+		path: norm, baseOff: baseOff, baseLen: baseLen,
+		depth: strings.Count(norm, "/"), flags: flags,
 	})
-
 	idx.byPath[norm] = id
-
 	for _, g := range extractTrigrams([]byte(norm)) {
 		idx.byGram[g] = append(idx.byGram[g], id)
 	}
-
 	if baseLen > 0 {
 		basename := []byte(norm[baseOff : baseOff+baseLen])
 		p1 := prefix1(basename)
@@ -82,11 +72,10 @@ func (idx *Index) Add(path string, flags uint16) uint32 {
 		p2 := prefix2(basename)
 		idx.byPrefix2[p2] = append(idx.byPrefix2[p2], id)
 	}
-
 	return id
 }
 
-// Remove marks the entry for path as deleted. Returns true if found.
+// Remove marks the entry for path as deleted.
 func (idx *Index) Remove(path string) bool {
 	norm := string(normalizePathBytes([]byte(path)))
 	id, ok := idx.byPath[norm]
@@ -106,8 +95,14 @@ func (idx *Index) Has(path string) bool {
 }
 
 // Len returns the number of live (non-deleted) documents.
-func (idx *Index) Len() int {
-	return len(idx.byPath)
+func (idx *Index) Len() int { return len(idx.byPath) }
+
+func copyPostings[K comparable](m map[K][]uint32) map[K][]uint32 {
+	cp := make(map[K][]uint32, len(m))
+	for k, v := range m {
+		cp[k] = v[:len(v):len(v)]
+	}
+	return cp
 }
 
 // Snapshot returns a frozen read-only view of the index.
@@ -116,30 +111,18 @@ func (idx *Index) Snapshot() *Snapshot {
 	for id := range idx.deleted {
 		del[id] = true
 	}
-
-	gramCopy := make(map[uint32][]uint32, len(idx.byGram))
-	for g, ids := range idx.byGram {
-		gramCopy[g] = ids[:len(ids):len(ids)]
-	}
-
 	var p1Copy [256][]uint32
 	for i, ids := range idx.byPrefix1 {
 		if len(ids) > 0 {
 			p1Copy[i] = ids[:len(ids):len(ids)]
 		}
 	}
-
-	p2Copy := make(map[uint16][]uint32, len(idx.byPrefix2))
-	for p, ids := range idx.byPrefix2 {
-		p2Copy[p] = ids[:len(ids):len(ids)]
-	}
-
 	return &Snapshot{
 		docs:      idx.docs[:len(idx.docs):len(idx.docs)],
 		count:     len(idx.docs),
 		deleted:   del,
-		byGram:    gramCopy,
+		byGram:    copyPostings(idx.byGram),
 		byPrefix1: p1Copy,
-		byPrefix2: p2Copy,
+		byPrefix2: copyPostings(idx.byPrefix2),
 	}
 }
