+++ agent/filefinder/text.go
@@ -1,5 +1,124 @@
 package filefinder
 
+import "slices"
+
+func toLowerASCII(b byte) byte {
+	if b >= 'A' && b <= 'Z' {
+		return b + ('a' - 'A')
+	}
+	return b
+}
+func normalizeQuery(q string) string {
+	b := make([]byte, 0, len(q))
+	prevSpace := true
+	for i := 0; i < len(q); i++ {
+		c := q[i]
+		if c == '\\' {
+			c = '/'
+		}
+		c = toLowerASCII(c)
+		if c == ' ' {
+			if prevSpace {
+				continue
+			}
+			prevSpace = true
+		} else {
+			prevSpace = false
+		}
+		b = append(b, c)
+	}
+	if len(b) > 0 && b[len(b)-1] == ' ' {
+		b = b[:len(b)-1]
+	}
+	return string(b)
+}
+func normalizePathBytes(p []byte) []byte {
+	j := 0
+	prevSlash := false
+	for i := 0; i < len(p); i++ {
+		c := p[i]
+		if c == '\\' {
+			c = '/'
+		}
+		c = toLowerASCII(c)
+		if c == '/' {
+			if prevSlash {
+				continue
+			}
+			prevSlash = true
+		} else {
+			prevSlash = false
+		}
+		p[j] = c
+		j++
+	}
+	return p[:j]
+}
+func extractTrigrams(s []byte) []uint32 {
+	if len(s) < 3 {
+		return nil
+	}
+	seen := make(map[uint32]struct{}, len(s))
+	for i := 0; i <= len(s)-3; i++ {
+		b0 := toLowerASCII(s[i])
+		b1 := toLowerASCII(s[i+1])
+		b2 := toLowerASCII(s[i+2])
+		gram := uint32(b0)<<16 | uint32(b1)<<8 | uint32(b2)
+		seen[gram] = struct{}{}
+	}
+	result := make([]uint32, 0, len(seen))
+	for g := range seen {
+		result = append(result, g)
+	}
+	slices.Sort(result)
+	return result
+}
+func extractBasename(path []byte) (offset int, length int) {
+	end := len(path)
+	if end > 0 && path[end-1] == '/' {
+		end--
+	}
+	if end == 0 {
+		return 0, 0
+	}
+	i := end - 1
+	for i >= 0 && path[i] != '/' {
+		i--
+	}
+	start := i + 1
+	return start, end - start
+}
+func extractSegments(path []byte) [][]byte {
+	var segments [][]byte
+	start := 0
+	for i := 0; i <= len(path); i++ {
+		if i == len(path) || path[i] == '/' {
+			if i > start {
+				segments = append(segments, path[start:i])
+			}
+			start = i + 1
+		}
+	}
+	return segments
+}
+func prefix1(name []byte) byte {
+	if len(name) == 0 {
+		return 0
+	}
+	return toLowerASCII(name[0])
+}
+func prefix2(name []byte) uint16 {
+	if len(name) == 0 {
+		return 0
+	}
+	hi := uint16(toLowerASCII(name[0])) << 8
+	if len(name) < 2 {
+		return hi
+	}
+	return hi | uint16(toLowerASCII(name[1]))
+}
+
+// ScoreParams controls the weights for each scoring signal.
 type ScoreParams struct {
 	BasenameMatch  float32
 	BasenamePrefix float32
@@ -23,7 +142,6 @@ func DefaultScoreParams() ScoreParams {
 		LengthPenalty:  0.01,
 	}
 }
-
 func isSubsequence(haystack, needle []byte) bool {
 	if len(needle) == 0 {
 		return true
@@ -39,7 +157,6 @@ func isSubsequence(haystack, needle []byte) bool {
 	}
 	return false
 }
-
 func longestContiguousMatch(haystack, needle []byte) int {
 	if len(needle) == 0 || len(haystack) == 0 {
 		return 0
@@ -68,11 +185,9 @@ func longestContiguousMatch(haystack, needle []byte) int {
 	}
 	return best
 }
-
 func isBoundary(b byte) bool {
 	return b == '/' || b == '.' || b == '_' || b == '-'
 }
-
 func countBoundaryHits(path []byte, query []byte) int {
 	if len(query) == 0 || len(path) == 0 {
 		return 0
@@ -88,7 +203,6 @@ func countBoundaryHits(path []byte, query []byte) int {
 	}
 	return hits
 }
-
 func equalFoldASCII(a, b []byte) bool {
 	if len(a) != len(b) {
 		return false
@@ -100,7 +214,6 @@ func equalFoldASCII(a, b []byte) bool {
 	}
 	return true
 }
-
 func hasPrefixFoldASCII(haystack, prefix []byte) bool {
 	if len(prefix) > len(haystack) {
 		return false
@@ -112,7 +225,6 @@ func hasPrefixFoldASCII(haystack, prefix []byte) bool {
 	}
 	return true
 }
-
 func scorePath(
 	path []byte,
 	baseOff int,
@@ -125,19 +237,14 @@ func scorePath(
 	if !isSubsequence(path, query) {
 		return 0
 	}
-
 	var score float32
-
 	basename := path[baseOff : baseOff+baseLen]
-
 	if isSubsequence(basename, query) {
 		score += params.BasenameMatch
 	}
-
 	if hasPrefixFoldASCII(basename, query) {
 		score += params.BasenamePrefix
 	}
-
 	segments := extractSegments(path)
 	for _, token := range queryTokens {
 		for _, seg := range segments {
@@ -147,15 +254,11 @@ func scorePath(
 			}
 		}
 	}
-
 	bh := countBoundaryHits(path, query)
 	score += float32(bh) * params.BoundaryHit
-
 	lcm := longestContiguousMatch(path, query)
 	score += float32(lcm) * params.ContiguousRun
-
 	score -= float32(depth) * params.DepthPenalty
 	score -= float32(len(path)) * params.LengthPenalty
-
 	return score
 }
